requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.1
pytz==2023.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
flask==2.3.2